# conecteai/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conecte.ai",
        "version": "1.0"
    },
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "ProductController"
                ],
                "summary": "Lista os produtos cadastrados.",
                "description": "Lista os produtos cadastrados no banco de dados.",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "$product"
                                ]
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductController"
                ],
                "summary": "Atualiza um produto.",
                "description": "Atualiza um novo produto no banco de dados.",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Galao de tinta"
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Azul"
                    },
                    {
                        "name": "dimensions",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "5 litros"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao atualizar produto."
                                ]
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Produto atualizado com sucesso."
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ProductController"
                ],
                "summary": "Cadastra um produto.",
                "description": "Cadastra um novo produto no banco de dados.",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Galao de tinta"
                    },
                    {
                        "name": "color",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Azul"
                    },
                    {
                        "name": "dimensions",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "5 litros"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao criar produto."
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao criar produto."
                                ]
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Produto criado efetuado com sucesso."
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "ProductController"
                ],
                "summary": "Mostra um produto cadastrado.",
                "description": "Mostra um produto cadastrado no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Produto não encontrado."
                                ]
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "$id"
                                ]
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductController"
                ],
                "summary": "Deleta um produto.",
                "description": "Deleta um produto existente no banco de dados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Erro ao deletar produto."
                                ]
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Produto não encontrado."
                                ]
                            }
                        }
                    },
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "example": [
                                    "Produto deletado com sucesso."
                                ]
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Faça login com email e senha para obter seu token.",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}